// intelligent-appointment/appointment-starter-kit/lib/models/IDateRangeOption.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.
import { IChoiceGroupOption } from '@fluentui/react';

export type DateRangeSelection = {
    startDate: Date;
    endDate: Date;
    range: string;
};

export interface IDateRangeOption extends IChoiceGroupOption {
    distance: number;
    calculateRange: (startDate?: Date, endDate?: Date) => DateRangeSelection;
}